+++ src/EncloseFieldTest.php
@@ -9,11 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace LeagueTest\Csv;
+namespace League\Csv;
 
 use InvalidArgumentException;
-use League\Csv\EncloseField;
-use League\Csv\Writer;
 use PHPUnit\Framework\TestCase;
 use function stream_get_filters;
 
@@ -21,14 +19,14 @@ use function stream_get_filters;
  * @group filter
  * @coversDefaultClass \League\Csv\EncloseField
  */
-class EncloseFieldTest extends TestCase
+final class EncloseFieldTest extends TestCase
 {
     /**
      * @see https://en.wikipedia.org/wiki/Comma-separated_values#Example
      *
      * @var array
      */
-    protected $records = [
+    private $records = [
             ['Year', 'Make', 'Model', 'Description', 'Price'],
             [1997, 'Ford', 'E350', 'ac,abs,moon', '3000.00'],
             [1999, 'Chevy', 'Venture "Extended Edition"', null, '4900.00'],